+++ lib/pages/dashboards/dashboards_list_page.dart
@@ -4,6 +4,7 @@ import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/services/nav_service.dart';
 import 'package:lotti/theme.dart';
+import 'package:lotti/widgets/charts/empty_dashboards_widget.dart';
 
 class DashboardsListPage extends StatefulWidget {
   const DashboardsListPage({Key? key}) : super(key: key);
@@ -36,6 +37,10 @@ class _DashboardsListPageState extends State<DashboardsListPage> {
                 dashboard.name.toLowerCase().contains(match) &&
                 dashboard.active)
             .toList();
+
+        if (items.isEmpty) {
+          return const EmptyDashboards();
+        }
 
         return ListView(
           shrinkWrap: true,