+++ linux/flutter/generated_plugins.cmake
@@ -5,6 +5,7 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   flutter_secure_storage_linux
   hotkey_manager
+  screen_retriever
   sqlite3_flutter_libs
   url_launcher_linux
   window_manager